wingify/requirements.txt
Django==1.9.5
argparse==1.2.1
decorator==4.0.9
django-cors-headers==1.1.0
djangorestframework==3.3.3
docutils==0.12
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
psycopg2==2.6.1
ptyprocess==0.5.1
requests==2.9.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
wsgiref==0.1.2
